[Change] move cookies and newsletter routes to system routes
