+++ index.js
@@ -100,11 +100,6 @@ app.use(
     })
 );
 
-mailchimp.setConfig({
-    apiKey: process.env.MAILCHIMP_API_KEY,
-    server: process.env.MAILCHIMP_SERVER_PREFIX
-});
-
 app.use("/api/books", checkDB, verifyClient, bookRoutes);
 app.use("/api/authors", checkDB, verifyClient, authorRoutes);
 app.use("/api/giftshop", checkDB, verifyClient, giftshopRoutes);
@@ -154,47 +149,6 @@ app.get("/api/logout", verifyClient, (req, res) => {
     res.end();
 });
 
-app.post("/api/client", (req, res) => {
-    const initialClient = req.body.headers["client-access-init"];
-
-    if (initialClient == "crisinus-client-net") {
-        const id = uuidv4();
-        const client = `crisinus-client-${id}`;
-        req.session.client = client;
-        const clientToken = jwt.sign({ id }, process.env.JWT_CLIENT_SECRET, {
-            expiresIn: 1000 * 60 * 60 * 24
-        });
-
-        res.cookie("client-access-token", clientToken, {
-            maxAge: 1000 * 60 * 60 * 24,
-            httpOnly: true
-        });
-
-        res.json({ clientReg: true, clientToken: clientToken, result: client });
-    } else {
-        res.send("API is for Crisinus client use only! xxx");
-    }
-});
-
-app.get("/api/cookiesmodal", (req, res) => {
-    res.cookie("cookies-modal", false, {
-        maxAge: 1000 * 60 * 60 * 24,
-        httpOnly: true
-    });
-
-    res.end();
-});
-
-app.get("/api/cookiesconfirm", verifyClient, (req, res) => {
-    const tempCookie = JSON.parse(req.cookies["cookies-modal"]);
-
-    if (typeof tempCookie == "boolean") {
-        res.json({ cookiesModal: tempCookie });
-    } else {
-        res.json({ cookiesModal: true });
-    }
-});
-
 app.post(
     "/api/login",
     checkDB,
@@ -235,47 +189,6 @@ app.post(
         } else {
             res.json({ auth: false, message: "No username/password match" });
             logger.info(`user ${username} failed to login`);
-        }
-    })
-);
-
-app.post(
-    "/api/newsletter",
-    checkDB,
-    verifyClient,
-    catchRequestError(async (req, res) => {
-        const email = req.body.email;
-
-        const [checkEmail] = await dbPoolPromise.execute(
-            "SELECT * FROM newsletter WHERE email = ?",
-            [email]
-        );
-
-        if (checkEmail[0]) {
-            res.send("You have already subscribed to our newsletter!");
-        } else {
-            const [saveEmail] = await dbPoolPromise.execute(
-                "INSERT INTO newsletter (email) VALUES (?)",
-                [email]
-            );
-
-            try {
-                const response = await mailchimp.lists.addListMember(
-                    process.env.MAILCHIMP_AUDIENCE_ID,
-                    {
-                        email_address: email,
-                        status: "subscribed",
-                        email_type: "html"
-                    }
-                );
-                res.send(
-                    `Thank you for your subscription on ${email}, ${response}`
-                );
-            } catch (err) {
-                console.log(err);
-                logger.error(`Mailchimp response fail: ${err}`);
-                res.status(400).send(err);
-            }
         }
     })
 );
